requirements.txt
fastapi
uvicorn
pandas
prophet
supabase
google-generativeai
python-multipart